Use SELinux context for insights-client unconditionally
This is nothing new and the SELinux context is already in use for
quite a long time. Running under this context does not require the
new SELinux policy introduced by the insights-core-selinux package.

Also remove tests as the complexity was greatly reduced and those
tests don't provide any reasonable value anymore.